Fixed the ability for someone not the currentUser to make changes or deletions on reviews not their own.

+++ app/controllers/reviews_controller.rb
@@ -17,7 +17,7 @@ class ReviewsController < ApplicationController
     
     def destroy
         review = Review.find_by(id: params[:id])
-        if review
+        if (review && review.user_id == session[:user_id])
             review.destroy
             render json: review, status: :ok
         else
@@ -27,7 +27,7 @@ class ReviewsController < ApplicationController
 
     def update
         review = Review.find_by(id: params[:id])
-        if review
+        if (review && review.user_id == session[:user_id])
             review.update!(review: params[:review]) 
             render json: review, execpt: :user, status: :ok
         else
